@@ -4,14 +4,15 @@ version: 0.0.1
 publish_to: "none"
 
 environment:
-  sdk: ">=3.0.1 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   cached_network_image: ^3.3.1
   flutter:
     sdk: flutter
   go_router: ^11.1.2
-  google_fonts: ^6.1.0
+  google_fonts: 6.1.0
   path_drawing: ^1.0.1
   shared_preferences: ^2.2.2
   zeta_flutter:
