@@ -1,5 +1,5 @@
 name: zeta_flutter
-version: 0.4.0
+version: 0.5.0
 description: Zeta is the new, formal, standardized Zebra Design System based off
   the successes of ZDS (Zebra Design System). This package is in pre-release,
   and so many aspects are incomplete.
@@ -17,8 +17,8 @@ platforms:
   linux:
 
 environment:
-  sdk: ">=3.0.1 <4.0.0"
-  flutter: ">=3.7.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
